requirements.txt
wheel
PyNaCl
discord.py
discord.py[voice]
pytz
jishaku
pytest
flake8
pylint
python-dotenv
chat-exporter